Added basic support for camera images
The images that come from the camera are encoded in YUV format, at
least in Android. The camera plugin returns them as an array of
bytes so they need special handling by this plugin. For now, we
added only a basic support with lots of steps that can be changed
to improve performance.